requirements.txt
# Note: CMake must be installed first:
# macOS: brew install cmake
# Ubuntu: apt-get install cmake
# Windows: Download from cmake.org

# Core dependencies
flask==2.2.5
requests==2.31.0
numpy>=1.24.3  # Updated to newer version
pandas>=2.0.3
scikit-learn>=1.0.2
matplotlib>=3.5.3
uvicorn>=0.23.2
websockets>=11.0.3

# ML/CV dependencies
ultralytics>=8.0.196
fastapi>=0.95.2
opencv-python-headless==4.10.0.84
mediapipe>=0.10.5
torch>=2.0.1
torchvision>=0.15.2

# Database and auth dependencies
mysql-connector-python>=8.0.33
SQLAlchemy>=2.0.23
python-multipart>=0.0.6
face-recognition>=1.3.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
email-validator>=2.0.0
roboflow>=1.1.58

# Configuration dependencies
python-dotenv>=1.0.0
pydantic-settings>=2.0.3
